backend/requirements.txt
# Django Framework (similar to Express.js in Node.js)
Django==4.2.7

# Django REST Framework (similar to Express + middleware for REST APIs)
djangorestframework==3.14.0

# PostgreSQL adapter (similar to 'pg' package in Node.js)
psycopg2-binary==2.9.9

# Database URL parsing for production (Railway/Render auto-provide DATABASE_URL)
dj-database-url==2.1.0

# Environment variables support (similar to 'dotenv' in Node.js)
python-decouple==3.8

# CORS headers (similar to 'cors' package in Node.js)
django-cors-headers==4.3.1

# Filtering support for querysets (similar to building WHERE clauses manually in SQL)
django-filter==23.5

# DRF GIS serializers for GeoJSON support
djangorestframework-gis==1.0

# For better development experience
django-extensions==3.2.3

# Timezone support
pytz==2023.3

# Production server (similar to 'pm2' or running Node with cluster mode)
gunicorn==21.2.0

# Whitenoise for serving static files (similar to express.static)
whitenoise==6.6.0
